Move checkForStaleOVSInterfaces and related code to node.go
While backporting to 4.12 the node healthz server, a number of functions related to checking stale ovs ports (checkForStaleOVSInternalPorts, checkForStaleOVSRepresentorInterfaces, checkForStaleOVSInterfaces) were moved to the pkg/node/openflow_manager.go and their related tests were left in pkg/node/healthcheck_test.go. In 4.13, we have everything under pkg/network-controller-manager.  To keep consistency, let's move these to pkg/node/node.go and pkg/node/node_test.go
